backend/requirements.txt
# Backend framework
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1

# Document parsing
pypdf==4.2.0
python-docx==1.1.2

# LangChain ecosystem (LOCKED & COMPATIBLE)
langchain==1.2.3
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0

# Vector database
faiss-cpu==1.8.0

# Tokenizer
tiktoken==0.7.0
